requirements.txt
runpod
pydantic-settings